DEV-24874 update move and wrap examples
